Fix dbus assert
I was getting this assert:

process 2: arguments to dbus_connection_ref() were incorrect, assertion "connection->generation == _dbus_current_generation" failed in file dbus-connection.c line 2686.
This is normally a bug in some application using the D-Bus library.
  D-Bus not built with -rdynamic so unable to print a backtrace

This is due to SDL calling dbus_shutdown() if it fails to open the
system bus, affecting all other users of dbus in the process (in
particular the Qt one).

To fix this we grant the app access to org.freedesktop.DBus which is
safe, but triggers the creation of a dbus proxy for the system bus.