Change filesystem permission to "rw"
As it is, there's no way to actually write to a save file, as the default launcher only gives the application read-only access to the filesystem. This attempts to fix that.

Keep in mind that I have no idea how flatpaks work or anything, so I'm not really sure if this is a proper fix. Manually running the application from the command line with `--filesystem=host:rw` added fixes the problem, so I'm inclined to believe it is.